requirements.txt
anthropic>=0.40.0
python-telegram-bot>=21.0
python-dotenv>=1.0.0
tzdata>=2024.1  # IANA timezone database, required on Windows
